NPC 처리 중간 저장. 처음 발견했을때 나한테 보내는걸 안함.

+++ 4차 과제/EchoServer/MMOSector.h
@@ -28,7 +28,7 @@ struct MMOSector
 
 ///////////////////
 	std::map<int16_t, MMOActor*>	actors;
-	std::set<int16_t>				npcs;
+	std::set<int32_t>				npcs;
 	MMONear*						near_sector_table[SECTOR_HEIGHT][SECTOR_WIDTH];
 	MMOZone*						zone;
 	int								index_x, index_y;
